функции расширения для преобразования модели вынесены их классов моделей
