+++ app/src/main/java/com/alexanderpodkopaev/androidacademyproject/data/ActorJsonModel.kt
@@ -10,14 +10,8 @@ data class ActorJsonModel(
     val id: Int,
     val name: String,
     @SerialName("profile_path")
-    val picture: String = ""
+    val picture: String?
 )
-
-fun ActorJsonModel.convertToModel(imageBaseUrl: String): Actor {
-    return Actor(
-        id = this.id, name = this.name, picture = imageBaseUrl + "original" + this.picture
-    )
-}
 
 @Keep
 @Serializable
