+++ app/src/main/java/com/alexanderpodkopaev/androidacademyproject/data/MovieJsonModel.kt
@@ -22,21 +22,3 @@ data class MovieJsonModel(
     @SerialName("vote_count")
     val voteCount: Int
 )
-
-fun MovieJsonModel.convertToModel(
-    imageBaseUrl: String
-): Movie {
-    return Movie(
-        id = this.id,
-        title = this.title,
-        overview = this.overview,
-        poster = imageBaseUrl + "original" + this.poster,
-        backdrop = imageBaseUrl + "original" + this.backdrop,
-        ratings = this.ratings,
-        adult = this.adult,
-        runtime = this.runtime,
-        actors = emptyList(),
-        genres = this.genres,
-        voteCount = this.voteCount
-    )
-}
