+++ app/src/main/java/com/alexanderpodkopaev/androidacademyproject/utils/ActorConverter.kt
@@ -1,0 +1,10 @@
+package com.alexanderpodkopaev.androidacademyproject.utils
+
+import com.alexanderpodkopaev.androidacademyproject.data.Actor
+import com.alexanderpodkopaev.androidacademyproject.data.ActorJsonModel
+
+fun ActorJsonModel.convertToModel(imageBaseUrl: String): Actor {
+    return Actor(
+        id = this.id, name = this.name, picture = imageBaseUrl + "original" + this.picture
+    )
+}
