+++ app/src/main/java/com/alexanderpodkopaev/androidacademyproject/utils/MovieConverter.kt
@@ -1,0 +1,22 @@
+package com.alexanderpodkopaev.androidacademyproject.utils
+
+import com.alexanderpodkopaev.androidacademyproject.data.Movie
+import com.alexanderpodkopaev.androidacademyproject.data.MovieJsonModel
+
+fun MovieJsonModel.convertToModel(
+    imageBaseUrl: String
+): Movie {
+    return Movie(
+        id = this.id,
+        title = this.title,
+        overview = this.overview,
+        poster = imageBaseUrl + "original" + this.poster,
+        backdrop = imageBaseUrl + "original" + this.backdrop,
+        ratings = this.ratings,
+        adult = this.adult,
+        runtime = this.runtime,
+        actors = emptyList(),
+        genres = this.genres,
+        voteCount = this.voteCount
+    )
+}
